Delete the inheritance relationship between Pattern and Operator
