+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/Operator.java
@@ -3,9 +3,14 @@ package com.starrocks.sql.optimizer.operator;
 
 import com.starrocks.sql.optimizer.OptExpression;
 import com.starrocks.sql.optimizer.OptExpressionVisitor;
+import com.starrocks.sql.optimizer.operator.scalar.ScalarOperator;
+
+import java.util.Objects;
 
 public abstract class Operator {
     protected final OperatorType opType;
+    protected long limit = -1;
+    protected ScalarOperator predicate;
 
     public Operator(OperatorType opType) {
         this.opType = opType;
@@ -19,30 +24,30 @@ public abstract class Operator {
         return false;
     }
 
-    public boolean isPattern() {
-        return false;
-    }
-
     public OperatorType getOpType() {
         return opType;
     }
 
-    @Override
-    public int hashCode() {
-        return opType.ordinal();
+    public long getLimit() {
+        return limit;
     }
 
-    @Override
-    public boolean equals(Object obj) {
-        if (!(obj instanceof Operator)) {
-            return false;
-        }
-        Operator rhs = (Operator) obj;
-        if (this == rhs) {
-            return true;
-        }
+    @Deprecated
+    public void setLimit(long limit) {
+        this.limit = limit;
+    }
 
-        return opType == rhs.getOpType();
+    public boolean hasLimit() {
+        return limit != -1;
+    }
+
+    public ScalarOperator getPredicate() {
+        return predicate;
+    }
+
+    @Deprecated
+    public void setPredicate(ScalarOperator predicate) {
+        this.predicate = predicate;
     }
 
     public <R, C> R accept(OperatorVisitor<R, C> visitor, C context) {
@@ -56,5 +61,23 @@ public abstract class Operator {
     @Override
     public String toString() {
         return opType.name();
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Operator operator = (Operator) o;
+        return limit == operator.limit && opType == operator.opType &&
+                Objects.equals(predicate, operator.predicate);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(opType.ordinal(), limit, predicate);
     }
 }
